+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 7.0.0 (2022-??-??)
 
 * Raised minimum required PHP version to 8.0.0
+* Added support for intersection types
 
 ## 6.2.1 (2020-12-18)